# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blossomer GTM API v2",
...
    "contact": {
      "name": "Blossomer GTM API Support",
      "email": "support@blossomer.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.blossomer.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/companies/": {
      "post": {
        "tags": [
          "Companies CRUD",
          "companies"
        ],
        "summary": "Create a new company",
        "description": "Create a new company record for the authenticated user with analysis data.",
        "operationId": "create_company_api_companies__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Company created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Company with this name already exists"
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "tags": [
          "Companies CRUD",
          "companies"
        ],
        "summary": "Get Companies",
        "description": "Get all companies for the authenticated user.\n\n- **skip**: Number of companies to skip (for pagination)\n- **limit**: Maximum number of companies to return (1-1000)",
        "operationId": "get_companies_api_companies__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of companies to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of companies to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of companies to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of companies to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyResponse"
                  },
                  "title": "Response Get Companies Api Companies  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/{company_id}": {
      "get": {
        "tags": [
          "Companies CRUD",
          "companies"
        ],
        "summary": "Get Company",
        "description": "Get a specific company by ID.\n\nOnly returns companies owned by the authenticated user.",
        "operationId": "get_company_api_companies__company_id__get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Companies CRUD",
          "companies"
        ],
        "summary": "Update Company",
        "description": "Update a company.\n\nOnly updates companies owned by the authenticated user.\nAll fields are optional - only provided fields will be updated.",
        "operationId": "update_company_api_companies__company_id__put",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Companies CRUD",
          "companies"
        ],
        "summary": "Delete Company",
        "description": "Delete a company and all related data.\n\nOnly deletes companies owned by the authenticated user.\nThis will cascade delete all accounts, personas, and campaigns.",
        "operationId": "delete_company_api_companies__company_id__delete",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/{company_id}/relations": {
      "get": {
        "tags": [
          "Companies CRUD",
          "companies"
        ],
        "summary": "Get Company With Relations",
        "description": "Get a company with all related accounts.\n\nOnly returns companies owned by the authenticated user.\nIncludes nested account data.",
        "operationId": "get_company_with_relations_api_companies__company_id__relations_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyWithRelations"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "account_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Account Data",
            "description": "Account data including firmographics, buying signals, rationale, metadata"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Company Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "account_data",
          "id",
          "company_id",
          "created_at",
          "updated_at"
        ],
        "title": "AccountResponse",
        "description": "Schema for account responses."
      },
...
      "CompanyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "url": {
            "type": "string",
            "maxLength": 500,
            "title": "Url"
          },
          "analysis_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Data"
          }
        },
        "type": "object",
        "required": [
          "name",
          "url"
        ],
        "title": "CompanyCreate",
        "description": "Schema for creating a new company.",
        "example": {
          "analysis_data": {
            "business_profile": {
              "business_model": "Monthly/annual subscriptions with tiered pricing",
              "category": "B2B SaaS workflow automation",
              "existing_customers": "50+ software companies using the platform"
            },
            "capabilities": [
              "Automated code review workflows",
              "CI/CD pipeline optimization",
              "Team collaboration tools",
              "Performance analytics dashboard"
            ],
            "description": "AI-powered workflow automation platform for software teams",
            "positioning": {
              "key_market_belief": "Manual dev processes are the biggest bottleneck in software delivery",
              "unique_approach": "AI-driven automation that learns from team patterns"
            }
          },
          "name": "TechFlow Solutions",
          "url": "https://techflowsolutions.com"
        }
      },
      "CompanyResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "url": {
            "type": "string",
            "maxLength": 500,
            "title": "Url"
          },
          "analysis_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Data"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid4",
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "url",
          "id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "CompanyResponse",
        "description": "Schema for company responses."
      },
      "CompanyUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "analysis_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Data"
          }
        },
        "type": "object",
        "title": "CompanyUpdate",
        "description": "Schema for updating company information.",
        "example": {
          "analysis_data": {
            "description": "Updated: AI-powered workflow automation platform for software teams",
            "last_updated": "2024-Q4"
          },
          "name": "TechFlow Solutions (Updated)"
        }
      },
      "CompanyWithRelations": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "url": {
            "type": "string",
            "maxLength": 500,
            "title": "Url"
          },
          "analysis_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Data"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid4",
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountResponse"
            },
            "type": "array",
            "title": "Accounts",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "url",
          "id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "CompanyWithRelations",
        "description": "Company schema with related accounts."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}